Add RootController to redirect root path to Swagger UI
Introduced a `RootController` to redirect requests from the root path (`/`) to the Swagger UI at `/index.html`. Additionally, updated `application-prod.properties` to include the `springdoc.swagger-ui.path` property and set `server.forward-headers-strategy` to `framework`.

The RootController is annotated as a Spring `@Controller` and enables cross-origin requests with `@CrossOrigin`. Its primary purpose is to handle requests to the root path and redirect them to the Swagger UI, improving API documentation accessibility. The property updates in `application-prod.properties` ensure proper configuration for Swagger UI and request forwarding in production.