Refactor replay and Player entities to support backlinks for efficient entity traversing.
Added concrete Player entity in database which is unique for each Player. Each Player contains multiple ReplayPlayer which are the player representation in the replays.

Added a PlayerSummary entity which is unique for each player and holds calculated summary information like rating history.

Completely refactored the PlayerController and PlayerRepository to support the new player structures.

Fixed the player ids in the ReplayPlayer to start from 0 instead of 1.

Refactored findSlice in ReplayRepositoryCustomImpl to implement the same logic in a SpliceHelper for reusability in the PlayerRepositoryCustomImpl.